First_Aid_buddy/requirements.txt
# ==============================================================================
# First-Aid Buddy Bot - Production Dependencies
# ==============================================================================
# Pinned versions for reproducible builds
# Last updated: 2025-12-07

# Web Framework
streamlit==1.29.0

# AI/ML
anthropic==0.8.1

# Configuration
python-dotenv==1.0.0

# ==============================================================================
# Development Dependencies (install separately)
# ==============================================================================
# pip install -r requirements-dev.txt

